refactor: imageFlow 정적 파일을 flowConfig.ts 하나로 병합
types.ts와 flowConfig.ts로 나뉘어 있던 정적 선언을 한 파일로 합칩니다.
둘을 합쳐도 170줄 규모라 분리의 이득(변경 이유 분리)보다 두 파일을 오가는 인지 비용이 컸습니다.

- types.ts 삭제, 내용을 flowConfig.ts로 이동 (도메인 값 → 데이터 타입 → 구조 타입 → 규칙 테이블 → 헬퍼 순)
- 값 표기 규칙 주석 추가 (외부에서 온 값은 대문자, 모듈 내부 구현 태그는 camelCase)
- importer 14곳 경로 갱신, 중복 import 3곳 병합
