+++ src/pages/banner/BannerDetailPage.tsx
@@ -14,7 +14,7 @@ import { useRecentFloorPlanQuery } from '@pages/imageSetup/v2/apis/queries/useRe
 
 import { ROUTES } from '@routes/paths';
 
-import { ENTRY_ROUTE } from '@store/imageFlow/types';
+import { ENTRY_ROUTE } from '@store/imageFlow/flowConfig';
 import { useImageFlowStore } from '@store/useImageFlowStore';
 
 import { GA_EVENTS } from '@shared/analytics/events';
