+++ src/pages/generate/v2/pages/result/utils/toProductItem.ts
@@ -1,4 +1,4 @@
-import type { ProductItem } from '@store/imageFlow/types';
+import type { ProductItem } from '@store/imageFlow/flowConfig';
 
 import type { GenerateImageResultProductResponse } from '@apis/__generated__/data-contracts';
