+++ src/shared/analytics/utils/imageEntryRoute/mapImageEntryRoute.ts
@@ -9,7 +9,7 @@
  *   image_entry_route: getEntryRoute(),
  * });
  */
-import { ENTRY_ROUTE, type EntryRoute } from '@store/imageFlow/types';
+import { ENTRY_ROUTE, type EntryRoute } from '@store/imageFlow/flowConfig';
 
 import {
   IMAGE_ENTRY_ROUTE,
